requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.40.0
timm>=0.9.7
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
Pillow>=10.0.0
clip @ git+https://github.com/openai/CLIP.git
streamlit>=1.27.0
tqdm>=4.66.0
matplotlib
gdown
huggingface_hub

# source fakenews_envi/Scripts/activate
# streamlit run app/streamlit_app.py
